pyproject.toml
[project]
name = "gpu-keeper"
version = "0.1.0"
description = "GPU utilization keeper - maintains GPU util at ~100% with automatic monitoring and restart"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch",
    "nvidia-ml-py",
    "pyyaml",
]

[project.scripts]
gpu-keeper = "gpu_keeper.main:run"
